backend/app/requirements/dev.txt
alembic==1.4.2
bcrypt==3.2.0
click==7.1.2
fastapi==0.61.0
h11==0.9.0
httptools==0.1.1
pydantic==1.6.1
SQLAlchemy==1.3.18
starlette==0.13.6
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
#psycopg2==2.8.5 - not working on m1 - dev in docker
pytest==6.0.1
requests==2.24.0
python-multipart==0.0.5
python-jose==3.2.0
passlib==1.7.2